+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # Changelog
 
-## next
+## 1.1.0
 
 - Add readyStates as both static and and instance members
 - default response `status` to 200
 - Pass mocked `ProgressEvent` and `MockEvent` to event listeners
+- Improve TypeScript typings
 
 ## 1.0.0